Moving Model to feature specified folder to lib/core/model
