+++ lib/app/core/model/activities/periksaklinikstatus_model.dart
