+++ lib/features/health_input/view/statuskunjunganselector_component.dart
@@ -1,5 +1,5 @@
 import 'package:afiyyah_connect/features/common/utils/extension/extensions.dart';
-import 'package:afiyyah_connect/features/health_input/model/periksaklinikstatus_model.dart';
+import 'package:afiyyah_connect/app/core/model/activities/periksaklinikstatus_model.dart';
 import 'package:flutter/material.dart';
 
 // enum PeriksaKlinikStatus { sudah, belum, luar }
